requirements.txt
asgiref==3.7.2
attrs==23.1.0
autobahn==23.6.1
Automat==22.10.0
captcha==0.5.0
certifi==2023.5.7
cffi==1.15.1
channels==4.0.0
channels-redis==4.0.0
charset-normalizer==3.1.0
constantly==15.1.0
cryptography==41.0.1
daphne==4.0.0
Django==4.1.3
django-anymail==10.2
django-cors-headers==3.13.0
django-ratelimit==4.1.0
django-sslserver==0.22
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
Faker==18.10.1
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
iniconfig==2.0.0
msgpack==1.0.5
packaging==23.1
Pillow==9.3.0
pluggy==1.0.0
psycopg2-binary==2.9.5
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
PyJWT==2.7.0
pyOpenSSL==23.2.0
pytest==7.3.1
pytest-asyncio==0.20.1
pytest-django==4.5.2
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
redis==4.5.5
requests==2.31.0
service-identity==21.1.0
six==1.16.0
sqlparse==0.4.4
stripe==5.4.0
Twisted==22.10.0
txaio==23.1.1
typing_extensions==4.6.3
urllib3==2.0.3
zope.interface==6.0
